aac: Import MediaTek changes

+++ libAACdec/src/channelinfo.cpp
@@ -1,4 +1,3 @@
-
 /* -----------------------------------------------------------------------------------------------------------
 Software License for The Fraunhofer FDK AAC Codec Library for Android
 
@@ -269,8 +268,8 @@ AAC_DECODER_ERROR getSamplingRateInfo(
     return AAC_DEC_UNSUPPORTED_FORMAT;
   }
 
-  FDK_ASSERT(t->ScaleFactorBands_Long[t->NumberOfScaleFactorBands_Long] == samplesPerFrame);
-  FDK_ASSERT(t->ScaleFactorBands_Short == NULL || t->ScaleFactorBands_Short[t->NumberOfScaleFactorBands_Short]*8 == samplesPerFrame);
+  FDK_ASSERT((UINT)(t->ScaleFactorBands_Long[t->NumberOfScaleFactorBands_Long]) == samplesPerFrame);
+  FDK_ASSERT(t->ScaleFactorBands_Short == NULL || (UINT)(t->ScaleFactorBands_Short[t->NumberOfScaleFactorBands_Short]*8) == samplesPerFrame);
 
   return AAC_DEC_OK;
 }
